fix: 해석기 운반·가림 판정을 남은 지점까지 마저 적용한다 (#94)
리뷰(R45) 세 건. 세 건 다 **직전 커밋들이 세운 규약의 미적용 지점**이다 —
규약은 맞게 잡았는데 적용 범위를 매번 좁게 잡았다.

① 재예약이 captured 해석기를 무시했다 (R45 #1)

R44 #1에서 `Input.sizeResolver`에 수집 시점 해석기를 실어 **배치**까지 들고
가게 했는데, 이월 각주를 새 페이지에서 다시 예약하는 `reservedFootnoteHeight`는
`input`을 순회하면서도 현재 `Environment`로 다시 쟀다. 그 사이 단·구역 기하가
바뀌면 배치(captured)와 재예약(현재)이 갈린다. `Environment.withSizeResolver`로
되돌려 재게 했다 — 수집→배치에 이어 수집→재예약도 닫힌다.

② 그림 가림이 클립을 무시했다 (R45 #2)

R43 #3에서 도형은 "`shapeCommands`가 경로만 칠하므로 경로 안쪽만 가린다"로
맞췄는데, 그림은 저작 rect 전체를 불투명으로 뒀다. `cellImageCommands`는
`clipRect`(rect와 같은 표-로컬 좌표) 안만 그리므로 페이지 절단면에 걸려 잘려
나간 부분은 아무것도 안 칠한다. 페인터와 같은 교집합(`rect ∩ clipRect`)으로
판정한다.

③ 가림이 블록 경계에서 소실됐다 (R45 #3)

`hyperlinkURL(for:)`이 `String?`라 `.occluded`가 nil로 접히고, `hit`의 frame 밖
경로가 `continue`로 **아래 블록으로 내려가** 각주 개체 밑에 깔린 링크를 열었다.
컨테이너 안에서는 가림을 지켰는데 블록 경계를 넘는 순간 잃었다. frame 밖 각주는
`containerHit`으로 직접 판정해 `.occluded`면 각주 히트를 돌려준다.

가드 (3건 전부 revert 시 실패 확인):
- `testOverflowReReservationUsesCapturedResolver` — 재예약 시점 기하가 바뀌어도
  결과 불변 + captured 값이 실제로 결과를 가른다(공허성 방지)
- `testClippedImageOccludesOnlyItsVisiblePart` — 보이는 절반은 가리고 잘린
  절반은 안 가린다(양쪽 단언)
- `testOcclusionOutsideBlockFrameDoesNotFallThroughToLowerBlock` — 아래에 깔린
  본문 링크 블록을 함께 두고 확인

`HwpFootnoteHyperlinkHitTests`가 file_length 상한(700)을 넘어 가림 계열을
`HwpFootnoteOcclusionHitTests.swift`로 분리했다 (294 + 447줄).

검증: `HWP_SNAPSHOT_TESTS=1 swift test` — 렌더 픽셀 해시가 **양 폰트 모드 모두**
전 픽스처 × 전 페이지 무변화(56.4s/59.5s 실제 실행), fidelity·블록 스냅샷·골든·
페이지 수 무변화. 전수 1,730개 0 실패.
