+++ Sources/HwpKitCore/AGENTS.md
@@ -135,7 +135,7 @@ CT 측정보다 우선한다 — 폰트 대체로 줄 수가 부풀어 배치가
 - `AnyHwpBlock.payload` — 종류별 상세 결과: `.table(HwpTableFrame)` / `.textbox` / `.footnote` / `.shape(HwpShapeGeometry)` / `.image(HwpImageBlockInfo)`. payload 내부 좌표는 **블록-로컬** (footnote 의 separator 만 페이지 좌표). 개체를 담는 컨테이너 payload는 셋이고 필드가 대칭이다 — `HwpTableCellFrame`(images/shapes/textboxes/nestedTables), `HwpTextboxFrame`(images/shapes), `HwpFootnoteBlock`(images/shapes/textboxes/nestedTables, #94)
 - `AnyHwpBlock.source` — `HwpBlockSource(controlInstanceId/paragraphId)`: 편집 기능이 렌더 결과에서 CoreHwp 모델로 돌아가는 참조
 - `AnyHwpBlock.hyperlinkURL` — block-level. `HwpPaginator.hyperlinkURL(in:)` 이 top-level 및 nested paragraph 양쪽에서 추출
-- 하이퍼링크 방출은 **스팬 우선**이다: `%hlk` 필드 스팬이 있으면 글리프 rect로만 방출하고, 스팬이 없는 컨테이너(표/글상자/각주) 문단만 문단 rect 폴백으로 담는다. 모델의 `hyperlinkURL`은 스팬 유무와 무관하게 채워지므로 폴백 측에서 `HwpDrawnTextLayout.hyperlinkRegions`로 게이트해야 앞뒤 평문이 링크로 표시되지 않는다 (hit tester의 `spanAwareHyperlinkURL`과 같은 규약). 컨테이너 **안쪽** 컨테이너까지 내려간다 — 각주 안 글상자·표 문단도 대상이라 방출은 `HwpPaintListBuilder.footnoteParagraphGroups` + `emitTable`이, 히트는 `HwpHitTester`의 `.footnote` 케이스가 **같은 깊이**를 걸어야 한다 (#94). 한쪽만 내려가면 밑줄은 그려지는데 탭이 안 먹거나 그 반대가 된다. 깊이만이 아니라 **영역**도 같아야 한다 — 각주 안 표는 블록 폭을 넘어 그려지는데 (한글도 안 자른다) `hit(page:point:)`의 블록 기각이 `block.frame`만 보면 그 띠의 링크가 안 눌린다. `HwpHitTester.hitEligibleFrame`이 각주 개체 rect까지 합집합으로 넓혀 페인트가 그리는 영역과 맞춘다 (R39 #3). 그 합집합과 순서는 **손으로 짜지 말고 walker에서 받는다** (R41): 자격 영역은 `walkFootnote` 방문 rect를 그대로 union하고 (최상위 rect만 모으면 각주 안 표의 셀·글상자 안 문단이 자기 컨테이너를 넘을 때 그 띠가 빠진다), 탐색은 **페인트 역순** — 안쪽 표 → 글 앞으로 개체 → 문단 텍스트 → 글 뒤로 개체 순이고 개체 정렬은 `footnoteTextboxesInPaintOrder`가 준다. 순서를 저장 순서로 두면 `.inFrontOfText` 글상자가 덮은 링크가 열린다. 컨테이너 rect로 미리 거르는 `where rect.contains` 게이트도 두지 않는다 — 포함 판정은 자손 rect를 아는 안쪽 함수 몫이다. **각주는 블록 전체 스팬 스캔(`walkText`)을 타지 않고** 그 층 인식 조회에 통째로 위임한다 (R42 #1) — `walkText`는 페인트 정순이라 덮인 문단 스팬을 먼저 잡고, 스팬 경로가 주 경로라 그대로 두면 역순 규약이 사실상 무효가 된다 (층 안에서도 문단마다 `spanAwareHyperlinkURL`이 스팬 우선을 지키므로 규칙은 잃지 않는다). 링크 없는 전경 층은 **불투명할 때만** 아래 탐색을 멈춘다 (`ContentLayer.occludes`, R42 #2) — 보이는 개체를 눌렀는데 숨은 링크가 열리면 안 되지만, 오버레이는 겹치는 것이 설계라 속 빈 장식 도형까지 막으면 그 아래 링크가 통째로 죽는다. **불투명 판정은 반드시 페인터를 보고 정한다** (R43): 글상자는 `fillColor`가 없어도 `textboxCommands`가 `.hwpWhite`로 칠하므로 **항상** 불투명하고("채우기 없음 = 투명"이 아니다), 도형은 `shapeCommands`가 `geometry.path`만 칠하므로 **경로 안쪽만** 가리며(바운딩 rect로 보면 타원 모서리의 링크가 죽는다), 표 셀은 `fillColor`가 있을 때 가린다. 그림만 알파를 알 수 없어 채워진 것으로 본다 — 그 한 가지가 한글.app 실측으로 확정할 몫이다. **컨테이너 rect로 미리 거르는 게이트는 어디에도 두지 않는다** — 중첩 표 rect(R41 #2)·셀 프레임(R44 #2)·글상자 rect(R44 #3) 셋 다, 자손이 자기 컨테이너를 넘어 그려질 수 있어 자격 영역은 인정하는데 조회만 막히기 때문이다. rect는 **가림 판정에만** 쓰고, 재귀 결과(`.occluded` 포함)는 **버리지 말고 그대로 전파**한다. 히트의 컨테이너 순회는 `containerHit` **하나**가 각주·표 셀·글상자에 재귀로 쓰인다 — 겹마다 따로 구현하면 다음 겹에서 또 갈린다(R39~R43이 전부 그 사례였다). 그래서 셀 글상자 전용 우회로(R31 #1)도 없앴다: 남겨 두면 가림으로 nil이 된 지점에서 덮인 링크가 되살아난다
+- 하이퍼링크 방출은 **스팬 우선**이다: `%hlk` 필드 스팬이 있으면 글리프 rect로만 방출하고, 스팬이 없는 컨테이너(표/글상자/각주) 문단만 문단 rect 폴백으로 담는다. 모델의 `hyperlinkURL`은 스팬 유무와 무관하게 채워지므로 폴백 측에서 `HwpDrawnTextLayout.hyperlinkRegions`로 게이트해야 앞뒤 평문이 링크로 표시되지 않는다 (hit tester의 `spanAwareHyperlinkURL`과 같은 규약). 컨테이너 **안쪽** 컨테이너까지 내려간다 — 각주 안 글상자·표 문단도 대상이라 방출은 `HwpPaintListBuilder.footnoteParagraphGroups` + `emitTable`이, 히트는 `HwpHitTester`의 `.footnote` 케이스가 **같은 깊이**를 걸어야 한다 (#94). 한쪽만 내려가면 밑줄은 그려지는데 탭이 안 먹거나 그 반대가 된다. 깊이만이 아니라 **영역**도 같아야 한다 — 각주 안 표는 블록 폭을 넘어 그려지는데 (한글도 안 자른다) `hit(page:point:)`의 블록 기각이 `block.frame`만 보면 그 띠의 링크가 안 눌린다. `HwpHitTester.hitEligibleFrame`이 각주 개체 rect까지 합집합으로 넓혀 페인트가 그리는 영역과 맞춘다 (R39 #3). 그 합집합과 순서는 **손으로 짜지 말고 walker에서 받는다** (R41): 자격 영역은 `walkFootnote` 방문 rect를 그대로 union하고 (최상위 rect만 모으면 각주 안 표의 셀·글상자 안 문단이 자기 컨테이너를 넘을 때 그 띠가 빠진다), 탐색은 **페인트 역순** — 안쪽 표 → 글 앞으로 개체 → 문단 텍스트 → 글 뒤로 개체 순이고 개체 정렬은 `footnoteTextboxesInPaintOrder`가 준다. 순서를 저장 순서로 두면 `.inFrontOfText` 글상자가 덮은 링크가 열린다. 컨테이너 rect로 미리 거르는 `where rect.contains` 게이트도 두지 않는다 — 포함 판정은 자손 rect를 아는 안쪽 함수 몫이다. **각주는 블록 전체 스팬 스캔(`walkText`)을 타지 않고** 그 층 인식 조회에 통째로 위임한다 (R42 #1) — `walkText`는 페인트 정순이라 덮인 문단 스팬을 먼저 잡고, 스팬 경로가 주 경로라 그대로 두면 역순 규약이 사실상 무효가 된다 (층 안에서도 문단마다 `spanAwareHyperlinkURL`이 스팬 우선을 지키므로 규칙은 잃지 않는다). 링크 없는 전경 층은 **불투명할 때만** 아래 탐색을 멈춘다 (`ContentLayer.occludes`, R42 #2) — 보이는 개체를 눌렀는데 숨은 링크가 열리면 안 되지만, 오버레이는 겹치는 것이 설계라 속 빈 장식 도형까지 막으면 그 아래 링크가 통째로 죽는다. **불투명 판정은 반드시 페인터를 보고 정한다** (R43): 글상자는 `fillColor`가 없어도 `textboxCommands`가 `.hwpWhite`로 칠하므로 **항상** 불투명하고("채우기 없음 = 투명"이 아니다), 도형은 `shapeCommands`가 `geometry.path`만 칠하므로 **경로 안쪽만** 가리며(바운딩 rect로 보면 타원 모서리의 링크가 죽는다), 표 셀은 `fillColor`가 있을 때 가린다. 그림은 알파를 알 수 없어 채워진 것으로 보되 **`clipRect` 안만** 가린다 (R45 #2 — `cellImageCommands`가 절단면 밖을 안 그린다). 알파 한 가지가 한글.app 실측으로 확정할 몫이다. 가림은 **블록 경계를 넘어서도** 유지된다 (R45 #3): `hit(page:point:)`가 frame 밖 각주를 `String?`로 받아 `.occluded`를 nil로 접으면 아래 블록으로 내려가 그 밑에 숨은 링크가 열린다. **컨테이너 rect로 미리 거르는 게이트는 어디에도 두지 않는다** — 중첩 표 rect(R41 #2)·셀 프레임(R44 #2)·글상자 rect(R44 #3) 셋 다, 자손이 자기 컨테이너를 넘어 그려질 수 있어 자격 영역은 인정하는데 조회만 막히기 때문이다. rect는 **가림 판정에만** 쓰고, 재귀 결과(`.occluded` 포함)는 **버리지 말고 그대로 전파**한다. 히트의 컨테이너 순회는 `containerHit` **하나**가 각주·표 셀·글상자에 재귀로 쓰인다 — 겹마다 따로 구현하면 다음 겹에서 또 갈린다(R39~R43이 전부 그 사례였다). 그래서 셀 글상자 전용 우회로(R31 #1)도 없앴다: 남겨 두면 가림으로 nil이 된 지점에서 덮인 링크가 되살아난다
 - **랜덤 UUID identifier 금지.** equality/hash 는 `frame + kind + text + url + payload + source` 기반. 같은 문서 두 번 로드 시 동일 블록으로 인식되어야 함
 - `HwpBlockKind`: `text` / `image` / `shape` / `table` / `textbox` / `footnote` / `placeholder`
 
@@ -374,7 +374,10 @@ CT 측정보다 우선한다 — 폰트 대체로 줄 수가 부풀어 배치가
   현재 단·문단 폭을 읽는 **계산 프로퍼티**라 시점마다 값이 다르므로, 예약이 쓴
   해석기를 `HwpFootnoteLayout.Input.sizeResolver`에 실어 배치까지 들고 간다.
   배치 시점에 다시 읽으면 그 사이 단 밴드가 바뀐 문서에서 `.column`·`.paragraph`
-  기준 개체가 다른 크기로 재조판된다 — 예약이 작으면 각주
+  기준 개체가 다른 크기로 재조판된다. **수집 시점 값을 쓰는 지점은 배치와
+  재예약 둘 다**다 (R45 #1) — 이월 각주를 새 페이지에서 다시 예약하는
+  `reservedFootnoteHeight`도 `input.sizeResolver`로 되돌려 재야 하고, 현재
+  environment로 재면 같은 각주를 배치와 다른 기하로 재게 된다 — 예약이 작으면 각주
   스택이 본문을 덮고, 크면 한글에 없는 페이지 절단이 생긴다. 코퍼스에는 떠 있는
   각주 개체가 0건이라 (전 픽스처 스캔: gso 1 + table 1, 둘 다 글자처럼 취급)
   이 하한은 헌법주석 페이지 수 1,030과 렌더 해시를 건드리지 않는다 —
